SPIKE: updated process outbox job to just trigger event
