+++ ghost/core/core/server/services/member-welcome-emails/index.js
@@ -1,4 +1,7 @@
 const jobs = require('./jobs');
+const StartMemberWelcomeEmailJobEvent = require('./events/StartMemberWelcomeEmailJobEvent');
+const domainEvents = require('@tryghost/domain-events');
+const processOutbox = require('./jobs/lib/process-outbox-main');
 
 class MemberWelcomeEmailsServiceWrapper {
     init() {
@@ -7,6 +10,13 @@ class MemberWelcomeEmailsServiceWrapper {
         }
 
         jobs.scheduleMemberWelcomeEmailJob();
+
+        // We currently cannot trigger a non-offloaded job from the job manager
+        // So the member welcome email job simply emits an event that we listen for here
+        // This allows the actual processing to run on the main thread instead of in a worker
+        domainEvents.subscribe(StartMemberWelcomeEmailJobEvent, async () => {
+            await processOutbox();
+        });
 
         this.initialized = true;
     }
