+++ ghost/core/core/server/services/member-welcome-emails/jobs/index.js
@@ -29,3 +29,4 @@ module.exports = {
         return hasScheduled.processOutbox;
     }
 };
+
